[jenkins] fix: check if multibranch job is active
problem: Active multibranch jobs are getting remove when the job cleanup code runs.
              Multibranch job cleanup logic remove active jobs that are not in the current branch job configuration.
              This is not correct since new projects that are not in dev branch will need their ci jobs.
solution: check if the multibranch job has any active jobs for any branch before deleting it.

+++ jenkins/shared-library/vars/createMonorepoMultibranchJobs.groovy
@@ -94,9 +94,18 @@ void removeOldJobs(String jobFolder, List<String> currentJobList) {
 	for (Item item in jobs) {
 		println("Checking Item remove: ${item.name}, ${item.getClass()} ${item.fullName}")
 		if (org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject == item.getClass()
-				&& currentJobList.indexOf(item.fullName) == -1) {
+				&& currentJobList.indexOf(item.fullName) == -1
+				&& !isJobActive(item)) {
 			println("Removing old job: ${item.fullName}")
 			item.delete()
 		}
 	}
+}
+
+boolean isJobActive(Item job) {
+	if (!job.isBuildable()) {
+		return false
+	}
+
+	return job.allJobs.any { item -> item.isBuildable() }
 }
